Move Download collection modals to the plugin
This patch moves all the download collection related modal components
from the eos-components package to the plugin.

https://phabricator.endlessm.com/T33744
