+++ kolibri_explore_plugin/assets/src/collections/intermediate-icons.png
