+++ kolibri_explore_plugin/assets/src/collections/primary-icons.png
