+++ kolibri_explore_plugin/assets/src/collections/secondary-icons.png
