+++ kolibri_explore_plugin/assets/src/components/CollectionSelectionModal.vue
