+++ kolibri_explore_plugin/assets/src/constants.js
@@ -43,3 +43,5 @@ export const searchTerms = new Map([
 
 export const COLLECTIONS_PAGE_SIZE = 300;
 export const SEARCH_MAX_RESULTS = 32;
+
+export const ApiURL = '/explore/api/endlesslearning';
