+++ packages/eos-components/src/constants.js
@@ -54,11 +54,8 @@ export const ItemsPerPage = 16;
 export const KeywordIconSize = 20;
 export const SearchBarDebounce = 2000;
 
-export const ApiURL = '/explore/api/endlesslearning';
-
 export default {
   AuthorFilterName,
-  ApiURL,
   CarouselInterval,
   DefaultKindLabel,
   FilterTagsBadList,
