+++ packages/eos-components/src/main.js
@@ -12,14 +12,11 @@ import ChannelCard from './components/ChannelCard.vue';
 import ChannelCardGroup from './components/ChannelCardGroup.vue';
 import ChannelLogo from './components/ChannelLogo.vue';
 import CollapsibleCardGrid from './components/CollapsibleCardGrid.vue';
-import CollectionSelectionModal from './components/CollectionSelectionModal.vue';
 import ContentCard from './components/ContentCard.vue';
 import ContentLink from './components/ContentLink.vue';
 import DummyButton from './components/DummyButton.vue';
 import Footer from './components/Footer.vue';
-import GradeSelectionModal from './components/GradeSelectionModal.vue';
 import GridPage from './components/GridPage.vue';
-import InstallContentModal from './components/InstallContentModal.vue';
 import Keywords from './components/Keywords.vue';
 import NavBar from './components/NavBar.vue';
 import PlayButton from './components/PlayButton.vue';
@@ -52,14 +49,11 @@ const components = {
   ChannelCardGroup,
   ChannelLogo,
   CollapsibleCardGrid,
-  CollectionSelectionModal,
   ContentCard,
   ContentLink,
   DummyButton,
   Footer,
-  GradeSelectionModal,
   GridPage,
-  InstallContentModal,
   Keywords,
   NavBar,
   PlayButton,
